Starting setting up configurable timesheet period
The idea would be to setup when you want a timesheet period to run until
and then the timesheet output would only return the timesheet for the
current period unless asked otherwise.

If the period can be calculated then we can build a running tally of
time spend over the period beyond just a single week.